refactor(querysvc): Move FindTraceSummaries tests to service_test.go
summary_test.go holds tests for helper functions (computeSummaries,
summarizeTrace); service_test.go holds tests for QueryService API
methods. Move the five TestFindTraceSummaries_* tests and their
supporting types (mockSummaryReader, wrappingReader) accordingly.

Also translate codes.Unimplemented to errors.ErrUnsupported in the
integration traceReader.FindTraceSummaries so QueryService can fall
back correctly when the remote server has not implemented the RPC.

Update ADR-010 to fold the ErrUnsupported design into section 5
(Fallback Logic) and tighten section 6 to describe only the gRPC
adapter's Unimplemented translation.
